created helper.py with load_data and filter_and_paginate functions pulled from app.py
